ref(gui): Move api calls to CSR(client side render) folder
