+++ frontend/app/register-device/RegisterDeviceForm.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Button from "@/components/Buttons/Button";
-import { getInfo, registerDevice } from "@/utils/api/auth";
+import { getInfo, registerDevice } from "@/utils/api/CSR/auth";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
